refactor(automation-runtime): own effect recovery and retry policy
