+++ crates/tracedecay-automation-runtime/src/automation/effect_runtime.rs
@@ -1,20 +1,27 @@
-//! Durable automation-effect journal, projection, retirement, and terminals.
-//!
-//! Daemon composition that needs the root `TraceDecay` aggregate (admission,
-//! retained-owner settlement, and crash-recovery orchestration) stays in the
-//! composition-root crate.
+//! Durable automation-effect journal, recovery, projection, and housekeeping.
 
+mod authority;
 pub mod contract;
 pub mod input;
 pub mod journal;
 pub mod problem;
 pub mod projection;
+#[cfg(any(test, feature = "test-helpers"))]
+pub mod recovery_index;
+#[cfg(not(any(test, feature = "test-helpers")))]
+mod recovery_index;
 pub mod retirement;
 pub mod terminal;
 
+pub use authority::finalize_terminal_housekeeping;
 pub use contract::{contract_error, digest};
 pub use input::{
     memory_curator_run_request, session_reflector_run_request, skill_writer_run_request,
     user_job_run_request,
+};
+pub use recovery_index::{
+    AutomationEffectRecoveryReport, add_pending_blocking, effect_authority_digest,
+    reconcile_reserved_automation_effects_for_project, recovered_partial_terminal,
+    remove_pending_blocking,
 };
 pub use terminal::{AutomationSettledProblem, AutomationSettledTerminal};
